+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `(no)space` and `(no)quote` manipulators to alter flags of the stream on the fly. - #129
+- Added `range` function to log a range of iterators. Can be used to log any custom container type as long as its value type can be logged. Can also be used to customize how a range/container is logged by specifying a custom formatting. - #137
 
 ## [4.4.1] - 2023-11-21
 ### General changes